+++ packages/cloud-cognitive/src/components/Notifications/Notifications.test.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2020, 2020
+ * Copyright IBM Corp. 2020, 2021
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -10,15 +10,19 @@ import { fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
 import '../../enable-all'; // must come before component is imported (directly or indirectly)
 import { pkg } from '../../settings';
+import uuidv4 from '../../global/js/utils/uuidv4';
 import { Notifications } from '.';
 
+const blockClass = `${pkg.prefix}--notifications-panel`;
+const dataTestId = uuidv4();
+
 describe('Notifications', () => {
-  test('renders the notification panel', () => {
+  it('renders the notification panel', () => {
     render(<Notifications data={[]} open setOpen={() => {}} />);
     expect(screen.queryAllByText(/Notifications/i)).toBeTruthy();
   });
 
-  test('should toggle do not disturb switch', () => {
+  it('should toggle do not disturb switch', () => {
     const { click } = fireEvent;
     const { fn } = jest;
     const onToggle = fn();
@@ -32,14 +36,12 @@ describe('Notifications', () => {
     );
 
     click(
-      container.querySelector(
-        `#${pkg.prefix}-notifications-panel-do-not-disturb-toggle-component`
-      )
+      container.querySelector(`#${blockClass}__do-not-disturb-toggle-component`)
     );
     expect(onToggle).toBeCalled();
   });
 
-  test('should render empty state illustration', () => {
+  it('should render empty state illustration', () => {
     const { container } = render(
       <Notifications data={[]} open setOpen={() => {}} />
     );
@@ -47,7 +49,7 @@ describe('Notifications', () => {
     expect(renderedEmptyStateSvg).toBeTruthy();
   });
 
-  test('should render notification with error state svg', () => {
+  it('should render notification with error state svg', () => {
     const { container } = render(
       <Notifications
         data={[
@@ -63,12 +65,12 @@ describe('Notifications', () => {
       />
     );
     const renderedEmptyStateSvg = container.querySelectorAll(
-      `svg.${pkg.prefix}-notifications-panel-notification-status-icon-error`
+      `svg.${blockClass}__notification-status-icon-error`
     );
     expect(renderedEmptyStateSvg[0]).toBeTruthy();
   });
 
-  test('should render notification with warning state svg', () => {
+  it('should render notification with warning state svg', () => {
     const { container } = render(
       <Notifications
         data={[
@@ -84,12 +86,12 @@ describe('Notifications', () => {
       />
     );
     const renderedEmptyStateSvg = container.querySelectorAll(
-      `svg.${pkg.prefix}-notifications-panel-notification-status-icon-warning`
+      `svg.${blockClass}__notification-status-icon-warning`
     );
     expect(renderedEmptyStateSvg[0]).toBeTruthy();
   });
 
-  test('should render notification with success state svg', () => {
+  it('should render notification with success state svg', () => {
     const { container } = render(
       <Notifications
         data={[
@@ -105,12 +107,12 @@ describe('Notifications', () => {
       />
     );
     const renderedEmptyStateSvg = container.querySelectorAll(
-      `svg.${pkg.prefix}-notifications-panel-notification-status-icon-success`
+      `svg.${blockClass}__notification-status-icon-success`
     );
     expect(renderedEmptyStateSvg[0]).toBeTruthy();
   });
 
-  test('should render notification with informational state svg', () => {
+  it('should render notification with informational state svg', () => {
     const { container } = render(
       <Notifications
         data={[
@@ -126,12 +128,12 @@ describe('Notifications', () => {
       />
     );
     const renderedEmptyStateSvg = container.querySelectorAll(
-      `svg.${pkg.prefix}-notifications-panel-notification-status-icon-informational`
+      `svg.${blockClass}__notification-status-icon-informational`
     );
     expect(renderedEmptyStateSvg[0]).toBeTruthy();
   });
 
-  test('should render link in notification', () => {
+  it('should render link in notification', () => {
     const { getByText, container } = render(
       <Notifications
         data={[
@@ -157,7 +159,7 @@ describe('Notifications', () => {
     expect(link.length && link).toEqual('https://www.carbondesignsystem.com/');
   });
 
-  test('should render Read more button', () => {
+  it('should render Read more button', () => {
     const { getByText } = render(
       <Notifications
         data={[
@@ -175,5 +177,29 @@ describe('Notifications', () => {
       />
     );
     expect(getByText(/Read more/i)).toBeTruthy();
+  });
+
+  it('adds additional properties to the containing node', () => {
+    const { container } = render(
+      <Notifications
+        open
+        setOpen={() => {}}
+        data={[]}
+        data-testid={dataTestId}
+      />
+    );
+    expect(
+      container.querySelector(
+        `.${blockClass}__container[data-testid="${dataTestId}"]`
+      )
+    ).toBeInTheDocument();
+  });
+
+  it('forwards a ref to an appropriate node', () => {
+    const ref = React.createRef();
+    render(<Notifications ref={ref} open setOpen={() => {}} data={[]} />);
+    expect(
+      ref.current.classList.contains(`${blockClass}__container`)
+    ).toBeTruthy();
   });
 });